requirements.txt
pyserial
pyfu-usb
intelhex
requests
semver
platformdirs
